[refactor] split GroupViewer into view and model
